requirements.txt
# Core ML Libraries
torch>=1.12.1,<2.0.0
transformers>=4.30.2,<5.0.0
huggingface-hub>=0.15.1
peft>=0.4.0

# Adversarial Attack Tools
nevergrad>=0.6.0
deap>=1.4.3

# Bioinformatics & Sequence Analysis
biopython>=1.81
Levenshtein>=0.20.9

# Data Processing & ML Utilities
numpy>=1.23.5
pandas>=1.5.3
scikit-learn>=1.1.3
tqdm>=4.64.1

# Visualization
matplotlib>=3.6.2
seaborn>=0.12.2

# Configuration & Logging
PyYAML>=6.0
mlflow>=2.0.0

# Development & Testing
pytest>=7.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
